requirements.txt
python-telegram-bot==20.2
python-dotenv==0.21.0
requests==2.27.1
